Update fingerprint_controller.rb
n the index method, instead of multiple separate queries, we now load all flagged fingerprints into a single hash (index_by(&:value)). This in-memory hash is used for lookups, which is much faster than hitting the database repeatedly.

User data is fetched in a single, optimized query.

N+1 Query Prevention: I've added Discourse.preloader.preload(users, :user_avatar) to load all user avatars at once, which is a standard and critical optimization in Discourse.

In user_report, User.find_by_username! is used. This will automatically raise an ActiveRecord::RecordNotFound error if the user doesn't exist, which Rails handles gracefully as a 404 Not Found error. This is more standard than raising Discourse::InvalidParameters.

In flag, save! is used to ensure that if a database save fails for any reason, it will raise an exception, making bugs easier to detect. destroy is used instead of delete to ensure any callbacks on the model are correctly triggered.